fix: prevent occasional generation of redirect page for paths conaining 'en' (#2306)
When docsUrl is set to an empty string, localization enabled and path to the file contains "en" letters, then pages with redirects would be generated instead of pages with contents.
So, it is impossible to reach the content.

Fixed: apply more strict rules of searching for locale part of url